backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.8
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
gunicorn==21.2.0